Codificacion de la respuesta del servidor

+++ BBDD/bbdd_para_import.sql
@@ -55,7 +55,7 @@ CREATE TABLE Statistics(
 	id int AUTO_INCREMENT,
 	id_user int,
 	id_game int,
-	timed int,
+	timed double,
 
 	PRIMARY KEY (id),
 	FOREIGN KEY (id_user) REFERENCES Users(id),
